Add directory refactor
