+++ src/authentication/LoginPage.js
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { useRef } from 'react';
 import { loginAsync } from '../redux/users/userSlice';
 import RegistrationForm from './RegistrationForm';
-import PasswordForm from './password';
+import PasswordForm from './passwords/forgotPassword';
 
 function LoginPage() {
   const formRef = useRef();
@@ -52,7 +52,7 @@ function LoginPage() {
           <div className="modal-content">
             <div className="modal-header">
               <h1 className="modal-title fs-5" id="staticBackdropLabel">Forgot your password?</h1>
-              <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close" />
+              <button type="button" className="btn-close reset-password" data-bs-dismiss="modal" aria-label="Close" />
             </div>
             <div className="modal-body">
               <PasswordForm />
@@ -70,7 +70,7 @@ function LoginPage() {
           <div className="modal-content">
             <div className="modal-header">
               <h1 className="modal-title fs-5" id="staticBackdropLabel">Sign up form</h1>
-              <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close" />
+              <button type="button" className="btn-close create-new" data-bs-dismiss="modal" aria-label="Close" />
             </div>
             <div className="modal-body">
               <RegistrationForm />
